--- src/platforms/native/StatusBar.js
@@ -1,17 +1,0 @@
-let exports = {};
-
-import { NATIVE } from 'base';
-
-exports.show = function () {
-  NATIVE.statusBar && NATIVE.statusBar.showStatusBar && NATIVE.statusBar.showStatusBar();
-};
-
-exports.hide = function () {
-  NATIVE.statusBar && NATIVE.statusBar.hideStatusBar && NATIVE.statusBar.hideStatusBar();
-};
-
-exports.getHeight = function () {
-  return NATIVE.statusBar && NATIVE.statusBar.statusBarHeight;
-};
-
-export default exports;
