+++ src/ui/Engine.js
@@ -48,18 +48,19 @@ import FPSView from 'ui/backend/debug/FPSView';
 
 import engineInstance from './engineInstance';
 
-import StackView from './StackView';
 import device from 'device';
 
 import InputEvent from 'event/input/InputEvent';
-import { getImport } from 'platformImport';
 
 import Matrix2D from '../platforms/browser/webgl/Matrix2D';
 
 var IDENTITY_MATRIX = new Matrix2D();
 
-const KeyListener = getImport('KeyListener');
-const InputListener = getImport('Input');
+import KeyListener from 'platforms/browser/KeyListener';
+import InputListener from 'platforms/browser/Input';
+
+import Canvas from 'platforms/browser/Canvas';
+import doc from 'platforms/browser/doc';
 
 var _timers = [];
 timer.onTick = function (dt) {
@@ -112,7 +113,6 @@ exports = class extends Emitter {
     this._doubleBuffered = true;
     this._countdown = null;
 
-    var Canvas = device.get('Canvas');
     this._rootElement = new Canvas({
       el: canvas,
       // use an existing canvas if one was provided, but wrap the 2D context
@@ -134,7 +134,7 @@ exports = class extends Emitter {
     this._ctx = this._rootElement.getContext('2d');
     this._ctx.font = '11px ' + device.defaultFontFamily;
 
-    this._view = opts.view || new StackView();
+    this._view = opts.view;
     this._view.style.update({
       width: opts.width,
       height: opts.height
@@ -170,7 +170,6 @@ exports = class extends Emitter {
       device.height = opts.height;
       device.screen.width = opts.width;
       device.screen.height = opts.height;
-      var doc = device.get('doc');
       if (doc) {
         doc.setEngine(this);
       }
