--- src/ui/backend/dom/Button.js
@@ -1,45 +1,0 @@
-let exports = {};
-
-/**
- * @license
- * This file is part of the Game Closure SDK.
- *
- * The Game Closure SDK is free software: you can redistribute it and/or modify
- * it under the terms of the Mozilla Public License v. 2.0 as published by Mozilla.
-
- * The Game Closure SDK is distributed in the hope that it will be useful,
- * but WITHOUT ANY WARRANTY; without even the implied warranty of
- * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE. See the
- * Mozilla Public License v. 2.0 for more details.
-
- * You should have received a copy of the Mozilla Public License v. 2.0
- * along with the Game Closure SDK.  If not, see <http://mozilla.org/MPL/2.0/>.
- */
-import { merge } from 'base';
-
-import TextView from 'ui/backend/dom/TextView';
-
-/**
- * @extends ui.backend.dom.TextView
- */
-exports = class extends TextView {
-  constructor (opts) {
-    opts = merge(opts, {
-      type: 'plain',
-      text: 'CHANGE ME',
-      color: 'white',
-      shadow: true,
-      'dom:className': 'button'
-    });
-
-    super(opts);
-
-    this._node.style.textShadow = 'black 2px 2px 2px';
-    this.setText(opts.text);
-  }
-  onInputSelect () {
-    this.publish('Select');
-  }
-};
-
-export default exports;
