+++ src/ui/keyboardTypes.js
@@ -3,7 +3,7 @@ let exports = {};
 import device from 'device';
 
 // map input types to native constants
-exports.nativeTypes = {
+exports.types = {
   // legacy native names
   Default: 0,
   // Default type for the current input method.
@@ -50,28 +50,17 @@ exports.htmlTypes = {
   month: 'month'
 };
 
-exports.types = {};
-
-for (var type in exports.nativeTypes) {
-  exports.types[type.toLowerCase()] = exports.nativeTypes[type];
+for (var type in exports.types) {
+  exports.types[type.toLowerCase()] = exports.types[type];
 }
-
-
-
 
 for (var type in exports.htmlTypes) {
   exports.htmlTypes[type.toLowerCase()] = exports.htmlTypes[type];
 }
 
 
-
-
 exports.has = function (type) {
   return type.toLowerCase() in exports.types;
-};
-
-exports.getNativeType = function (type) {
-  return exports.types[type.toLowerCase()] || 0;
 };
 
 exports.getHTMLType = function (type) {
