+++ src/ui/resource/Image.js
@@ -28,7 +28,6 @@ import {
   GLOBAL,
   bind,
   CONFIG,
-  NATIVE,
   logger
 } from 'base';
 
@@ -37,6 +36,7 @@ import PubSub from 'lib/PubSub';
 import Callback from 'event/Callback';
 import resourceLoader from 'ui/resource/loader';
 import filterRenderer from 'ui/backend/canvas/filterRenderer';
+import Canvas from 'platforms/browser/Canvas';
 
 var ImageCache = {};
 
@@ -73,27 +73,20 @@ resourceLoader.on(resourceLoader.IMAGE_LOADED, function (image, src) {
 /**
  * This class models the region of a larger image that this "Image" references.
  */
-var ImageMap = !CONFIG.disableNativeViews && NATIVE.timestep && NATIVE.timestep
-  .ImageMap;
-if (!ImageMap) {
-  ImageMap = class {
-    constructor (parentImage, x, y, width, height, marginTop, marginRight,
-      marginBottom, marginLeft, url) {
-      this.url = url;
-      this.x = x;
-      this.y = y;
-      this.width = width;
-      this.height = height;
-      this.marginTop = marginTop;
-      this.marginRight = marginRight;
-      this.marginBottom = marginBottom;
-      this.marginLeft = marginLeft;
-    }
-  };
-}
-
-var isNative = GLOBAL.NATIVE && !device.isNativeSimulator;
-var Canvas = device.get('Canvas');
+class ImageMap {
+  constructor (parentImage, x, y, width, height, marginTop, marginRight,
+    marginBottom, marginLeft, url) {
+    this.url = url;
+    this.x = x;
+    this.y = y;
+    this.width = width;
+    this.height = height;
+    this.marginTop = marginTop;
+    this.marginRight = marginRight;
+    this.marginBottom = marginBottom;
+    this.marginLeft = marginLeft;
+  }
+};
 
 // helper canvases for image data, initialized when/if needed
 var _imgDataCanvas = null;
@@ -140,18 +133,6 @@ exports = class extends PubSub {
         img = b64;
       } else if (b64) {
         url = b64;
-      }
-    }
-
-    if (forceReload) {
-      // clear native texture in an image object
-      if (img && img.destroy) {
-        img.destroy();
-      }
-
-      // clear native textures by URL
-      if (url && NATIVE.gl && NATIVE.gl.deleteTexture) {
-        NATIVE.gl.deleteTexture(url);
       }
     }
 
@@ -206,24 +187,7 @@ exports = class extends PubSub {
         cb = cb.chain();
       }
 
-      var hasFired = this._cb.fired();
-
-      // GC native has a reload method to force reload
-      if (srcImg.reload) {
-        var onReload = bind(this, function () {
-          srcImg.removeEventListener('reload', onReload, false);
-          if (hasFired) {
-            this._cb.fire(null, this);
-          }
-          cb && cb();
-        });
-        srcImg.addEventListener('reload', onReload, false);
-        srcImg.reload();
-
-        if (hasFired) {
-          this._cb.reset();
-        }
-      } else if (cb) {
+      if (cb) {
         if (this._cb.fired()) {
           // always wait a frame before calling the callback
           setTimeout(cb, 0);
@@ -400,7 +364,7 @@ exports = class extends PubSub {
 
     var srcImg = this._srcImg;
 
-    if (!isNative && ctx.filter) {
+    if (ctx.filter) {
       var filterImg = filterRenderer.renderFilter(ctx, this, srcX, srcY, srcW, srcH);
       if (filterImg) {
         srcImg = filterImg;
